ai_literature_assistant/requirements.txt
# ai_literature_assistant/requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pypdf>=3.17.0
streamlit>=1.28.0
langchain-text-splitters>=0.0.1